Enable extra pedantic warnings and fix resulting issues

+++ console.c
@@ -238,7 +238,7 @@ consoleread(struct inode *ip, char *dst, int n)
     }
     c = input.buf[input.r++ % INPUT_BUF];
     if(c == C('D')){  // EOF
-      if(n < target){
+      if((uint)n < target){
         // Save ^D for next time, to make sure
         // caller gets a 0-byte result.
         input.r--;
@@ -253,7 +253,7 @@ consoleread(struct inode *ip, char *dst, int n)
   release(&input.lock);
   ilock(ip);
 
-  return target - n;
+  return (int)(target - (uint)n);
 }
 
 int
